refactor: discount.js 분리
