+++ src/basic/utils/cart.js
@@ -1,5 +1,3 @@
-const TOTAL_BULK_DISCOUNT_RATE = 0.25;
-
 export const products = [
   { id: 'p1', name: '상품1', price: 10000, quantity: 50 },
   { id: 'p2', name: '상품2', price: 20000, quantity: 30 },
@@ -7,64 +5,6 @@ export const products = [
   { id: 'p4', name: '상품4', price: 15000, quantity: 0 },
   { id: 'p5', name: '상품5', price: 25000, quantity: 10 },
 ];
-
-export const calculateDiscountedPrice = (price, discountRate) => price * (1 - discountRate);
-const calculateDiscountRate = (totalPrice, discountedTotalPrice) => (totalPrice - discountedTotalPrice) / totalPrice;
-
-export const getProductBulkDiscountRate = (productId, quantity) => {
-  const PRODUCT_BULK_DISCOUNT_AMOUNT = 10;
-  const PRODUCT_BULK_DISCOUNT_RATE = {
-    p1: 0.1,
-    p2: 0.15,
-    p3: 0.2,
-    p4: 0.05,
-    p5: 0.25,
-  };
-
-  if (quantity >= PRODUCT_BULK_DISCOUNT_AMOUNT) return PRODUCT_BULK_DISCOUNT_RATE[productId];
-  return 0;
-};
-
-const getMoreDiscountPriceAndRate = (discountedTotalPrice, totalPrice) => {
-  let updatedTotalPrice = 0;
-  let discountRate = 0;
-
-  const bulkDiscountingPrice = discountedTotalPrice * TOTAL_BULK_DISCOUNT_RATE;
-  const itemBulkDiscountingPrice = totalPrice - discountedTotalPrice;
-
-  if (bulkDiscountingPrice > itemBulkDiscountingPrice) {
-    updatedTotalPrice = calculateDiscountedPrice(totalPrice, TOTAL_BULK_DISCOUNT_RATE);
-    discountRate = TOTAL_BULK_DISCOUNT_RATE;
-  } else {
-    updatedTotalPrice = discountedTotalPrice;
-    discountRate = calculateDiscountRate(totalPrice, discountedTotalPrice);
-  }
-
-  return { updatedTotalPrice, discountRate };
-};
-export const calculateTotalProductsBulkDiscount = (totalItems, totalPrice, discountedTotalPrice) => {
-  const TOTAL_BULK_DISCOUNT_AMOUNT = 30;
-
-  if (totalItems < TOTAL_BULK_DISCOUNT_AMOUNT) {
-    return {
-      updatedTotalPrice: discountedTotalPrice,
-      discountRate: calculateDiscountRate(totalPrice, discountedTotalPrice),
-    };
-  }
-
-  return getMoreDiscountPriceAndRate(discountedTotalPrice, totalPrice);
-};
-export const calculateDayDiscount = ({ updatedTotalPrice, discountRate }) => {
-  const SALE_DAY = 2;
-  const SALE_DAY_DISCOUNT_RATE = 0.1;
-
-  if (new Date().getDay() === SALE_DAY) {
-    updatedTotalPrice = calculateDiscountedPrice(updatedTotalPrice, SALE_DAY_DISCOUNT_RATE);
-    discountRate = Math.max(discountRate, SALE_DAY_DISCOUNT_RATE);
-  }
-
-  return { updatedTotalPrice, discountRate };
-};
 
 export const updateTotalInfo = (discountedTotalPrice, discountRate) => {
   const cartTotalInfo = document.getElementById('cart-total');
@@ -94,14 +34,19 @@ export const updateBonusPoints = (bonusPoints, totalPrice) => {
 
   return updatedBonusPoints;
 };
-export const renderProductsStockInfo = () => {
-  const stockInfo = document.getElementById('stock-status');
-  let infoMsg = '';
 
-  products.forEach((item) => {
-    if (item.quantity < 5) {
-      infoMsg += item.name + ': ' + (item.quantity > 0 ? '재고 부족 (' + item.quantity + '개 남음)' : '품절') + '\n';
-    }
-  });
-  stockInfo.textContent = infoMsg;
+const formatLowStocksInfo = (lowStockProducts) => {
+  return lowStockProducts
+    .map((product) =>
+      product.quantity > 0 ? `${product.name}: 재고 부족 (${product.quantity}개 남음)` : `${product.name}: 품절`,
+    )
+    .join('\n');
+};
+export const renderProductsStockInfo = () => {
+  const LACK_OF_STOCK = 5;
+
+  const stockInfo = document.getElementById('stock-status');
+  const lowStockProducts = products.filter((product) => product.quantity < LACK_OF_STOCK);
+
+  stockInfo.textContent = formatLowStocksInfo(lowStockProducts);
 };
